Rearrange classes, add async update support
