+++ java/src/main/java/VaultClient.java
@@ -6,13 +6,18 @@ import com.google.gson.reflect.TypeToken;
 import javax.naming.AuthenticationException;
 import java.io.IOException;
 import java.lang.reflect.Type;
+import java.net.URI;
 import java.net.URLEncoder;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
 import java.nio.charset.StandardCharsets;
+import java.time.Duration;
 import java.util.*;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-public class VaultClient extends VaultClientMethods {
+public class VaultClient {
   private final Logger logger = Logger.getLogger(VaultClient.class.getName());
   private String vaultAddr;
   private String vaultToken;
@@ -190,22 +195,25 @@ public class VaultClient extends VaultClientMethods {
     VaultSecret secret = this.secrets.get(path);
     if (!secret.leased) {
       // No lease yet
-      getSecret(path);
+      getSecret(secret);
     } else {
       secondsSinceSecretLease = (currentTime.getTime() - secret.leaseTime.getTime()) / 1000;
       if (secondsSinceSecretLease >= secret.leaseDuration) {
         // Lease expired
-        getSecret(path);
+        getSecret(secret);
       } else if ((float) secondsSinceSecretLease > ((float) secret.leaseDuration * (2.0 / 3.0))) {
         if (secret.renewable) {
           // Lease is renewable
           if (!secret.updateLock) {
             secret.updateLock = true;
-            updateSecret(secret);
+            renewSecret(secret);
           }
         } else {
           // Lease is not renewable
-          getSecret(path);
+          if (!secret.updateLock) {
+            secret.updateLock = true;
+            updateSecret(secret);
+          }
         }
       }
     }
@@ -218,15 +226,14 @@ public class VaultClient extends VaultClientMethods {
     }
   }
 
-  private void getSecret(String path) throws IOException, InterruptedException {
+  private void getSecret(VaultSecret secret) throws IOException, InterruptedException {
     String secretJson;
-    VaultSecret secret = this.secrets.get(path);
     Map<String, String> vaultHeaders = new HashMap<>();
     Gson gson = new Gson();
 
     vaultHeaders.put("X-Vault-Token", this.vaultToken);
 
-    secretJson = this.httpGet(this.vaultAddr + "/v1" + path, vaultHeaders);
+    secretJson = this.httpGet(this.vaultAddr + "/v1" + secret.path, vaultHeaders);
     if (secretJson != null) {
       secret.leaseTime = new Date();
       secret.value = gson.fromJson(secretJson, JsonObject.class);
@@ -249,11 +256,16 @@ public class VaultClient extends VaultClientMethods {
     }
   }
 
+  private void renewSecret(VaultSecret secret) {
+    VaultSecretRenewRunnable renewer = new VaultSecretRenewRunnable();
+    renewer.setSecret(secret);
+
+    new Thread(renewer).start();
+  }
+
   private void updateSecret(VaultSecret secret) {
     VaultSecretUpdateRunnable updater = new VaultSecretUpdateRunnable();
     updater.setSecret(secret);
-    updater.setVaultAddr(this.vaultAddr);
-    updater.setVaultToken(this.vaultToken);
 
     new Thread(updater).start();
   }
@@ -291,7 +303,7 @@ public class VaultClient extends VaultClientMethods {
     String loginResultJson;
     Type listOfStrings = new TypeToken<ArrayList<String>>() {}.getType();
     final String loginPath = String.format("/v1/auth/%s/login", this.authPath);
-    VaultLoginData loginData = new VaultLoginData();
+    VaultLoginJwtData loginData = new VaultLoginJwtData();
 
     Gson gson = new Gson();
 
@@ -323,5 +335,159 @@ public class VaultClient extends VaultClientMethods {
       return true;
     }
     return false;
+  }
+
+  public JsonElement getDottedJsonPath(String path, JsonObject object) {
+    List<String> pathList = Arrays.asList(path.split("\\.").clone());
+    return getListJsonPath(pathList, object);
+  }
+
+  public JsonElement getListJsonPath(List<String> path, JsonObject object) {
+    JsonElement result = object.get(path.get(0));
+    if (result == null) {
+      return null;
+    }
+    if (path.size() > 1) {
+      return getListJsonPath(path.subList(1, path.size()), result.getAsJsonObject());
+    } else {
+      return result;
+    }
+  }
+
+  public String httpGet(String uri, Map<String, String> headers)
+      throws IOException, InterruptedException {
+    return httpRequest("GET", uri, headers, null);
+  }
+
+  // With the Vault API, currently POST and PUT are interchangeable methods
+  public String httpPost(String uri, String data) throws IOException, InterruptedException {
+    return httpRequest("POST", uri, new HashMap<>(), data);
+  }
+
+  public String httpPost(String uri, Map<String, String> headers, String data)
+      throws IOException, InterruptedException {
+    return httpRequest("POST", uri, headers, data);
+  }
+
+  private String httpRequest(String method, String uri, Map<String, String> headers, String data)
+      throws IOException, InterruptedException {
+    HttpClient client = HttpClient.newHttpClient();
+    // TODO: Allow timeout to be configurable
+    HttpRequest.Builder requestBuilder =
+        HttpRequest.newBuilder().timeout(Duration.ofSeconds(3)).uri(URI.create(uri));
+
+    if (data == null) {
+      requestBuilder = requestBuilder.method(method, HttpRequest.BodyPublishers.noBody());
+    } else {
+      requestBuilder = requestBuilder.method(method, HttpRequest.BodyPublishers.ofString(data));
+    }
+
+    for (Map.Entry<String, String> entry : headers.entrySet())
+      requestBuilder.header(entry.getKey(), entry.getValue());
+
+    HttpRequest request = requestBuilder.build();
+    HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
+
+    if (response.statusCode() > 200 && response.statusCode() < 400) {
+      logger.warning(
+          "Unexpected response. The Vault client is not configured to handle redirects: "
+              + response.headers()
+              + " : "
+              + response.body());
+      return null;
+    } else if (response.statusCode() >= 400) {
+      logger.severe("Error occurred: " + response.body());
+      return null;
+    }
+
+    return response.body();
+  }
+
+  public static class VaultSecret {
+    public String path;
+    public JsonObject value;
+    public String leaseId;
+    public Date leaseTime;
+    public long leaseDuration;
+    public boolean leased = false;
+    public boolean renewable = false;
+    public boolean updateLock = false;
+
+    public VaultSecret(String path) {
+      this.path = path;
+    }
+  }
+
+  private static class VaultLoginJwtData {
+    public String role;
+    public String jwt;
+  }
+
+  private static class VaultRenewalData {
+    public String lease_id;
+    public long increment;
+  }
+
+  private class VaultSecretUpdateRunnable implements Runnable {
+    private VaultSecret secret;
+
+    public void setSecret(VaultSecret secret) {
+      this.secret = secret;
+    }
+
+    public void run() {
+
+      try {
+        VaultClient.this.getSecret(this.secret);
+      } catch (IOException | InterruptedException e) {
+        e.printStackTrace();
+      }
+
+      this.secret.updateLock = false;
+    }
+  }
+
+  private class VaultSecretRenewRunnable implements Runnable {
+    private VaultSecret secret;
+
+    public void setSecret(VaultSecret secret) {
+      this.secret = secret;
+    }
+
+    public void run() {
+
+      Map<String, String> vaultHeaders = new HashMap<>();
+      String renewalResultJson = null;
+      String renewalDataJson;
+      VaultRenewalData renewalData = new VaultRenewalData();
+      Gson gson = new Gson();
+
+      vaultHeaders.put("X-Vault-Token", VaultClient.this.vaultToken);
+
+      renewalData.lease_id = this.secret.leaseId;
+      renewalData.increment = this.secret.leaseDuration;
+      renewalDataJson = gson.toJson(renewalData);
+
+      try {
+        renewalResultJson =
+            VaultClient.this.httpPost(
+                VaultClient.this.vaultAddr + "/v1/sys/leases/renew", vaultHeaders, renewalDataJson);
+      } catch (IOException | InterruptedException e) {
+        e.printStackTrace();
+      }
+
+      JsonObject renewalResult = gson.fromJson(renewalResultJson, JsonObject.class);
+      if (renewalResult != null
+          && VaultClient.this.getDottedJsonPath("lease_id", renewalResult) != null) {
+        secret.leaseTime = new Date();
+        secret.leaseId =
+            VaultClient.this.getDottedJsonPath("lease_id", renewalResult).getAsString();
+        secret.renewable =
+            VaultClient.this.getDottedJsonPath("renewable", renewalResult).getAsBoolean();
+        secret.leaseDuration =
+            VaultClient.this.getDottedJsonPath("lease_duration", renewalResult).getAsLong();
+      }
+      secret.updateLock = false;
+    }
   }
 }
