--- java/src/main/java/VaultLoginData.java
@@ -1,4 +1,0 @@
-public class VaultLoginData {
-  public String role;
-  public String jwt;
-}
