--- java/src/main/java/VaultRenewalData.java
@@ -1,4 +1,0 @@
-public class VaultRenewalData {
-  public String lease_id;
-  public long increment;
-}
